+++ apps/remnant2toolkit/app/navbar.tsx
@@ -27,7 +27,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
       desktopChildren={
         <>
           <Menu as="div" className="relative">
-            <Menu.Button className="bg-background text-surface-solid hover:text-primary-500 flex text-sm font-semibold">
+            <Menu.Button className="flex text-sm font-semibold bg-background text-surface-solid hover:text-primary-500">
               {({ active }) => (
                 <div className="flex items-center justify-center gap-1">
                   <span className="absolute -inset-1.5" />
@@ -68,7 +68,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
                       )}
                     >
                       <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.createBuild.icon className="text-primary-600 h-5 w-5" />
+                        <NAV_ITEMS.createBuild.icon className="w-5 h-5 text-primary-600" />
                       </div>
                       <div className="flex flex-col items-start justify-start gap-y-1">
                         {NAV_ITEMS.createBuild.label}
@@ -89,7 +89,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
                       )}
                     >
                       <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.featuredBuilds.icon className="text-primary-600 h-5 w-5" />
+                        <NAV_ITEMS.featuredBuilds.icon className="w-5 h-5 text-primary-600" />
                       </div>
                       <div className="flex flex-col items-start justify-start gap-y-1">
                         {NAV_ITEMS.featuredBuilds.label}
@@ -110,33 +110,12 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
                       )}
                     >
                       <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.communityBuilds.icon className="text-primary-600 h-5 w-5" />
+                        <NAV_ITEMS.communityBuilds.icon className="w-5 h-5 text-primary-600" />
                       </div>
                       <div className="flex flex-col items-start justify-start gap-y-1">
                         {NAV_ITEMS.communityBuilds.label}
                         <p className="text-xs font-normal text-gray-400">
                           {NAV_ITEMS.communityBuilds.description}
-                        </p>
-                      </div>
-                    </BaseLink>
-                  )}
-                </Menu.Item>
-                <Menu.Item>
-                  {({ active }) => (
-                    <BaseLink
-                      href={NAV_ITEMS.collectionBuilds.href}
-                      className={cn(
-                        active ? 'bg-gray-800' : '',
-                        'text-surface-solid flex w-full flex-row items-start justify-start p-2 text-sm font-semibold',
-                      )}
-                    >
-                      <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.collectionBuilds.icon className="text-primary-600 h-5 w-5" />
-                      </div>
-                      <div className="flex flex-col items-start justify-start gap-y-1">
-                        {NAV_ITEMS.collectionBuilds.label}
-                        <p className="text-xs font-normal text-gray-400">
-                          {NAV_ITEMS.collectionBuilds.description}
                         </p>
                       </div>
                     </BaseLink>
@@ -152,7 +131,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
                       )}
                     >
                       <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.gimmickBuilds.icon className="text-primary-600 h-5 w-5" />
+                        <NAV_ITEMS.gimmickBuilds.icon className="w-5 h-5 text-primary-600" />
                       </div>
                       <div className="flex flex-col items-start justify-start gap-y-1">
                         {NAV_ITEMS.gimmickBuilds.label}
@@ -173,7 +152,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
                       )}
                     >
                       <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.baseGameBuilds.icon className="text-primary-600 h-5 w-5" />
+                        <NAV_ITEMS.baseGameBuilds.icon className="w-5 h-5 text-primary-600" />
                       </div>
                       <div className="flex flex-col items-start justify-start gap-y-1">
                         {NAV_ITEMS.baseGameBuilds.label}
@@ -194,7 +173,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
                       )}
                     >
                       <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.beginnerBuilds.icon className="text-primary-600 h-5 w-5" />
+                        <NAV_ITEMS.beginnerBuilds.icon className="w-5 h-5 text-primary-600" />
                       </div>
                       <div className="flex flex-col items-start justify-start gap-y-1">
                         {NAV_ITEMS.beginnerBuilds.label}
@@ -215,7 +194,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
                       )}
                     >
                       <div className="mr-4 w-[20px]">
-                        <NAV_ITEMS.randomBuild.icon className="text-primary-600 h-5 w-5" />
+                        <NAV_ITEMS.randomBuild.icon className="w-5 h-5 text-primary-600" />
                       </div>
                       <div className="flex flex-col items-start justify-start gap-y-1">
                         {NAV_ITEMS.randomBuild.label}
@@ -282,7 +261,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.createBuild.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -299,7 +278,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.featuredBuilds.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -316,7 +295,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.communityBuilds.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -329,28 +308,11 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
           </BaseLink>
 
           <BaseLink
-            href={NAV_ITEMS.collectionBuilds.href}
-            className="flex flex-row items-center justify-start"
-          >
-            <NAV_ITEMS.collectionBuilds.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
-              aria-hidden="true"
-            />
-            <div className="flex flex-col items-start justify-start px-3 py-2">
-              {NAV_ITEMS.collectionBuilds.label}
-
-              <p className="text-xs text-gray-400">
-                {NAV_ITEMS.collectionBuilds.description}
-              </p>
-            </div>
-          </BaseLink>
-
-          <BaseLink
             href={NAV_ITEMS.gimmickBuilds.href}
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.gimmickBuilds.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -367,7 +329,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.baseGameBuilds.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -384,7 +346,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.beginnerBuilds.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -401,7 +363,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.randomBuild.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -420,7 +382,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.itemLookup.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -437,7 +399,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.itemTracker.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -454,7 +416,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.itemQuiz.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -471,7 +433,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.resources.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">
@@ -490,7 +452,7 @@ export function Navbar({ showNotifications }: { showNotifications: boolean }) {
             className="flex flex-row items-center justify-start"
           >
             <NAV_ITEMS.supportR2TK.icon
-              className="text-primary-600 mr-2 h-7 w-5 flex-none"
+              className="flex-none w-5 mr-2 text-primary-600 h-7"
               aria-hidden="true"
             />
             <div className="flex flex-col items-start justify-start px-3 py-2">